Move useIsDateBeyondRetentionPeriod to my-sites/backup/hooks
