+++ client/my-sites/backup/hooks.js
@@ -1,12 +1,17 @@
 /**
  * External dependencies
  */
-import { useEffect, useRef } from 'react';
+import { isEnabled } from '@automattic/calypso-config';
+import { useCallback, useEffect, useRef } from 'react';
 import { useSelector } from 'react-redux';
 
 /**
  * Internal dependencies
  */
+import { applySiteOffset } from 'calypso/lib/site/timezone';
+import getSiteActivityLogRetentionDays from 'calypso/state/selectors/get-site-activity-log-retention-days';
+import getSiteGmtOffset from 'calypso/state/selectors/get-site-gmt-offset';
+import getSiteTimezoneValue from 'calypso/state/selectors/get-site-timezone-value';
 import { getHttpData } from 'calypso/state/data-layer/http-data';
 import { getRequestActivityLogsId, requestActivityLogs } from 'calypso/state/data-getters';
 import { requestRewindCapabilities } from 'calypso/state/rewind/capabilities/actions';
@@ -141,4 +146,35 @@ export const useFirstMatchingBackupAttempt = (
 		isLoading: isLoadingActivityLogs,
 		backupAttempt: matchingAttempt || undefined,
 	};
+};
+
+/**
+ * A React hook that creates a callback to test whether or not a given date is
+ * within a site's Backup retention period (if retention periods are enabled).
+ *
+ * @param {number|null} siteId The site whose retention period we'll be testing against.
+ * @returns A callback that returns true if a given date is outside the site's retention period, and false otherwise.
+ */
+export const useIsDateBeyondRetentionPeriod = ( siteId ) => {
+	const gmtOffset = useSelector( ( state ) => getSiteGmtOffset( state, siteId ) );
+	const timezone = useSelector( ( state ) => getSiteTimezoneValue( state, siteId ) );
+	const retentionDays = useSelector( ( state ) =>
+		getSiteActivityLogRetentionDays( state, siteId )
+	);
+
+	return useCallback(
+		( date ) => {
+			if ( ! isEnabled( 'activity-log/retention-policies' ) ) {
+				return false;
+			}
+
+			if ( retentionDays === undefined ) {
+				return false;
+			}
+
+			const today = applySiteOffset( Date.now(), { gmtOffset, timezone } ).startOf( 'day' );
+			return today.diff( date, 'days' ) > retentionDays;
+		},
+		[ gmtOffset, timezone, retentionDays ]
+	);
 };
